added optional static input mapping, prepared for runtime encoding and scaling

+++ pythonapp/main.py
@@ -1,20 +1,23 @@
 import torch
 import pandas
 import models
+from sklearn.preprocessing import LabelEncoder, StandardScaler
 
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
 keyData = pandas.read_csv("kb.csv", header=None)
-keyTensor = torch.tensor(keyData.values, dtype=torch.float64).to(device)
+keyTensor = torch.tensor(keyData.values, dtype=torch.float32).to(device)
 print(keyTensor)
 
 moveData = pandas.read_csv("mmove.csv", header=None)
-moveTensor = torch.tensor(moveData.values, dtype=torch.float64).to(device)
+moveTensor = torch.tensor(moveData.values, dtype=torch.float32).to(device)
 print(moveTensor)
 
 clickData = pandas.read_csv("mclick.csv", header=None)
-clickTensor = torch.tensor(clickData.values, dtype=torch.float64).to(device)
+clickTensor = torch.tensor(clickData.values, dtype=torch.float32).to(device)
 print(clickTensor)
+
+#Preprocess
 
 keySplit = int(0.8 * len(keyTensor))
 moveSplit = int(0.8 * len(moveTensor))
